parser-requirements.txt
# Document parsing libraries for POC testing
markitdown[all]>=0.1.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
pdfminer.six>=20221105
pypdf>=3.0.0
unstructured[all-docs]>=0.10.0

# Additional dependencies
python-magic>=0.4.27
pillow>=10.0.0
